feat(stepper): convert to typescript
